Bump PHPStan to level 5 and fix all findings (222 -> 0)
Level 5 adds argument-type checking and dead-code detection
(always-true/false conditions, unreachable branches, impossible
instanceof checks). Fixed via dispatched per-file clusters plus a
personal mop-up pass; each finding was root-caused as either a real
bug or a too-narrow type declaration before fixing.

Real bugs found and fixed:
- Criteria::merge() used `==` instead of `=` when resetting
  defaultCombineOperator, silently no-op'ing the reset.
- ObjectBuilder::addColumnEnum() passed a caught exception as the
  `$code` (int) param of EngineException instead of `$previous`,
  silently dropping the exception chain.
- ObjectBuilder's addClearAllReferences() guarded generated code with
  `if ($deep = true)` (assignment, not comparison).
- XmlToAppData never applied its own $encoding to xml_parser_create().
- Column::setTypeFromString() wrote to a dead $size property instead
  of $domain->setSize(), silently discarding the value.
- Table::toXML() passed a bool to DOMElement::setAttribute() where a
  string is required for isCrossRef.
- PropulsionAutoloader::register()/unregister() used different callable
  shapes, so unregister() could never actually remove the callback.
- PropulsionXMLParser::convertDOMElementToArray() let non-element
  child nodes (comments, CDATA) leak into the output array.
- ModelCriteria's find/count/delete/deleteAll/update paths assigned
  Propulsion::getConnection()'s PDO|PropulsionPDO result straight into
  PropulsionPDO-typed parameters; added instanceof guards (matching
  the existing PropulsionCollection pattern) instead of risking a
  native TypeError when a non-PropulsionPDO connection class is
  configured.
- ModelCriteria::getCountStatement() caught only PropulsionException,
  unlike its sibling getSelectStatement(), so PDO failures during
  count-query prep bypassed the friendly-wrapping and cleanup logic;
  widened to catch Exception.
- Restored PropulsionException's instanceof-Exception single-arg
  calling convention, which a prior automated pass had dropped as
  "unreachable" without noticing PropulsionExceptionTest exercises it.

Everything else was a stale/too-narrow docblock type (nullable
properties documented as non-nullable, e.g. Table/Database's
baseClass/basePeer/interface, ScopedElement's namespace/schema/package,
Domain's size/scale) corrected to match real runtime behavior, plus a
handful of genuinely dead code/properties/constants removed.

Full 2263-test suite passes clean after the fixes.

+++ runtime/Lib/Util/PropulsionDateTime.php
@@ -56,9 +56,9 @@ class PropulsionDateTime extends DateTime
 			return true;
 		}
 
-		$month = date('m', $stamp);
-		$day   = date('d', $stamp);
-		$year  = date('Y', $stamp);
+		$month = (int) date('m', $stamp);
+		$day   = (int) date('d', $stamp);
+		$year  = (int) date('Y', $stamp);
 
 		return checkdate($month, $day, $year);
 	}
