+++ runtime/Lib/Util/PropulsionModelPager.php
@@ -74,7 +74,7 @@ class PropulsionModelPager implements \IteratorAggregate, \Countable
 		if (($this->getPage() == 0 || $this->getMaxPerPage() == 0)) {
 			$this->setLastPage(0);
 		} else {
-			$this->setLastPage(ceil($this->getNbResults() / $this->getMaxPerPage()));
+			$this->setLastPage((int) ceil($this->getNbResults() / $this->getMaxPerPage()));
 
 			$offset = ($this->getPage() - 1) * $this->getMaxPerPage();
 			$q->offset($offset);
